Modified components folder to rename files to PascalCase
