+++ app/client/src/components/JoinButton.js
